fix: made placeholders case-insensitive and renamed it so it matches the naming scheme
